+++ include/CQChartsPlotType.h
@@ -73,6 +73,7 @@ class CQChartsPlotType : public QObject {
   Q_PROPERTY(bool      isGroupType        READ isGroupType       )
   Q_PROPERTY(bool      isHierarchical     READ isHierarchical    )
   Q_PROPERTY(bool      canProbe           READ canProbe          )
+  Q_PROPERTY(bool      canRectSelect      READ canRectSelect     )
 
   Q_ENUMS(Dimension)
 
@@ -130,6 +131,8 @@ class CQChartsPlotType : public QObject {
   virtual bool isHierarchical() const { return false; }
 
   virtual bool canProbe() const = 0;
+
+  virtual bool canRectSelect() const { return true; }
 
   //---